+++ export_academy/models.py
@@ -58,7 +58,7 @@ class Event(TimeStampedModel, ClusterableModel):
         ordering = ('-created', '-modified')
 
     def __str__(self):
-        return f'{self.id}: {self.name}'
+        return f'{self.id}:{self.name}'
 
 
 class Registration(TimeStampedModel):
