+++ pii-secret-exclude.txt
@@ -68,3 +68,5 @@ international_online_offer/static/fonts/EuclidFlexB-Medium.otf
 international_online_offer/sass/components/_footer.scss
 international_online_offer/templates/ioo/base.html
 styles/colors/colors.scss
+tests/unit/export_academy/factories.py
+tests/unit/export_academy/test_views.py
